Adjust Somnia to be deployable
