+++ apps/somnia/package.json
@@ -1,0 +1,7 @@
+{
+  "name": "somnia",
+  "version": "0.0.1",
+  "scripts": {
+    "start": "node main.js"
+  }
+}
