+++ apps/somnia/src/main.ts
@@ -37,14 +37,15 @@ client.on(Events.MessageReactionAdd, (...args) =>
 //collect commands
 import * as archive from './commands/archive'
 import * as button from './commands/button'
-import * as databank from './commands/databank'
+// TODO: Enable after i18n of databank
+// import * as databank from './commands/databank'
 export const Commands: Collection<string, any> = new Collection()
 Commands.set(archive.slashcommand.name, archive)
-Commands.set(databank.slashcommand.name, databank)
+// Commands.set(databank.slashcommand.name, databank)
 Commands.set(button.slashcommand.name, button)
 const setcommands = [
   archive.slashcommand.toJSON(),
-  databank.slashcommand.toJSON(),
+  // databank.slashcommand.toJSON(),
   button.slashcommand.toJSON(),
 ]
