+++ libs/gi/assets-data/executors.json
@@ -1,0 +1,9 @@
+{
+  "executors": {
+    "gen-assets-data": {
+      "implementation": "./src/executors/gen-assets-data/executor",
+      "schema": "./src/executors/gen-assets-data/schema.json",
+      "description": "gen-assets-data executor"
+    }
+  }
+}
