+++ libs/gi/assets-data/package.json
@@ -1,0 +1,6 @@
+{
+  "name": "@genshin-optimizer/gi/assets-data",
+  "version": "0.0.1",
+  "type": "commonjs",
+  "executors": "./executors.json"
+}
