+++ libs/gi/assets-data/src/executors/gen-assets-data/schema.d.ts
@@ -1,0 +1,1 @@
+export interface GenAssetsDataExecutorSchema {} // eslint-disable-line
