+++ libs/sr/assets-data/src/AssetsData_gen.json
