+++ libs/sr/assets-data/src/executors/gen-assets-data/schema.json
@@ -1,0 +1,9 @@
+{
+  "$schema": "https://json-schema.org/schema",
+  "version": 2,
+  "title": "GenAssetsData executor",
+  "description": "",
+  "type": "object",
+  "properties": {},
+  "required": []
+}
