+++ libs/sr/assets-data/src/index.ts
@@ -1,1 +1,4 @@
-export * from './AssetData'
+import * as AssetData_gen from './AssetsData_gen.json'
+import type { AssetData as AssetDataType } from './executors/gen-assets-data/executor'
+
+export const AssetData = AssetData_gen as AssetDataType
