--- libs/sr/assets-data/src/lib/util.ts
@@ -1,4 +1,0 @@
-export function trimPath(path: string) {
-  if (!path) path = ''
-  return path.split('/').at(-1) ?? ''
-}
